Allow power.py to use intel_pstate driver
While we don't get the level of control of the acpi_cpufreq driver when
using the intel_pstate driver, we can still cap the max_freq. So this
patch allows use of the intel_pstate driver, but warns the used when
certain options are not available.
Also, because we need to query MSRs to find the base frequency, warn the
user then the msr kernel module is not loaded, and ask them to load it.

Signed-off-by: David Hunt <david.hunt@intel.com>