add extra check for when no high priority cores
When an SST-BF enabled BIOS is used with an incompatible kernel, this
may result in all frequencies in the base_frequency sysfs files to
show the same frequency. In this case, we will notify the user, ask
them to confirm that a compatible BIOS and kernel are in use, then
exit.

Signed-off-by: David Hunt <david.hunt@intel.com>