pwr: fix SST-BF detection
Currently, we unconditionally set SST-BF to be enabled when we
detect base_frequency sysfs entries. This is not actually how
we should check for SST-BF, because these sysfs entries are
there unconditionally after a certain kernel version.

What actually indicates that SST-BF is enabled in the BIOS is
that 1) sysfs path for base frequency exists in the first place,
and 2) that different cores contain different values for that.

Therefore, remove all SST-BF detection in per-core code, and
move it to after we've gathered all the data about all the
cores.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>