pwr: add support for detecting whether EPP is enabled
Currently, the script will throw an exception if EPP is disabled
in the BIOS. Add support for ignoring EPP at runtime, as well as
provide a flag for applications to check whether EPP is enabled.

EPP is detected with a CPUID bit, but we cannot rely on user
having CPUID installed, so we have to parse /proc/cpuinfo instead.
So, we're adding an internal-only "cpuinfo" module to parse the
/proc/cpuinfo information.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>