pwr: fix negative time check
When we're checking for negative time differences to account for
clock drift/changes on a system that lacks a monotonic clock,
we're checking it as a new condition, not a continuation of
previous check. This results in the value of prev_ts being
None at the time of check if previously we have reset prev_ts
due to it being longer than 60 seconds since last power
consumption read.

Fix it to be an alternate branch of previous check rather than
a standalone check.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>