fix error message when SST-BF not enabled
Error message wrongly says that base_frequency file not found in
sysfs when the contents of that file equals P1 frequency. It's far
more likely that SST-BF is not enabled in the BIOS, resulting in
all the base_frequency files showing P1 rather than P1 High and P1
normal frequencies.

Signed-off-by: David Hunt <david.hunt@intel.com>