pwr: ensure all frequencies are valid before testing configuration
Currently, we only test core configuration for validity only in case of
SST-BF enabled. This breaks when non-SST-BF configurations are used when
max frequency is lower than all-core turbo frequency. Fix the check to be
unconditional and catch any instances of desired max being higher than
available max. This will ensure that if the frequency is higher than P1,
it is either valid for all configurations, or invalid for all
configurations.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>