fix frequency range to work with acpi-cpufreq
When the acpi-cpufreq driver is being used, all turbo frequencies are
show as the P1 frequency plus 1, i.e. 2201. When generating the list
of valid frequencies with the previous algorithm, the 2201 would appear
in the list as 2300. This fix generates the list, and inserts entries
in 100MHz increments up to just below the max, then adds the max. So when
the max freq is 2200, yout get ... 2100 2200, and when the max is 2201,
you get ... 2100 2200 2201 as expected.

Signed-off-by: David Hunt <david.hunt@intel.com>