ansible: multi action lowpower provisioning (#3)
This change is introducing the deploy-powerupdown.yml
provisioning script which will take care of setting the
system to below profiles.

1.lowpower:
This profile tuns C6 C-state "on/off" for the list of cores the user
wants.
* User has to specify the list of cores and C6 C-state on which
  C6 C-state should be turned "on/off".

2. very-lowpower:
This profile tuns C6 C-state "on/off" for the list of cores the user
wants. Also lowers the max frequency of the cores by the given number
of bins.
* User will specify the list of cores on which C6 C-state should be
  "on/off"
* User will also specify the number of bins by which max frequency
  of the listed cores should be lowered.

The user should specify the target hosts as groups under the file
* /etc/ansible/hosts
  (or)
* /etc/ansible/hosts.yaml

The user should specify the core list, C6 C-state, bins and action details
as variables in the below files
* /etc/ansible/group_vars/<serversgroup.yml>
  example:
    /etc/ansible/group_vars/webservers.yaml

The action sections with variables should be added like below.
Note: One can add multiple action under "actions_to_apply"

actions_to_apply:
- action: lowpower
  c6state: 'on'
  core_list: 10,11,14,15,92,93
- action: very-lowpower
  bins_to_lower: 2
  c6state: 'off'
  core_list:
- action: very-lowpower
  bins_to_lower: 4
  c6state: 'off'
  core_list: 8,9,12,13,16,17

Note: Do not add actions with empty core list.