power.py improvements (#9)
* power.py: add error handling for sysfs file opening

When passing a core range with cores beyond the last core on the system,
the script would just crash. This commit adds exception handling so that
the error is gracefully handled and the user can continue configuring
the systems power states without relaunching the script.

Signed-off-by: Kevin Laatz <kevin.laatz@intel.com>

* power.py: rename parameter in set_cstate

The path used for enabling/disabling the cstate via this function is
"{sysfs_path}/disable". The parameter passed to the function in order to
set the desired state in called "enable". This makes the code confusing to
read, so change the parameter name to "disable" for consistency and
clarity.

Signed-off-by: Kevin Laatz <kevin.laatz@intel.com>

* power.py: fix help text example

One of the examples in the help text had a mismatch between the comment
and example command. This commit fixes this.

Signed-off-by: Kevin Laatz <kevin.laatz@intel.com>