power.py: imporove err handling for uncore, turbo
* power.py: cleanup uncore frequency handling

This commit makes the following change to the uncore frequency
configuration code.

1) The uncore frequency is changed for all the cores on the system.
So, move the repeated duplicate code of getting full cpu range to
set_uncore_frequency() function.

* power.py: fix exception for no turbo file

When intel_pstate driver is disabled on the system,
the turbo file /sys/devices/system/cpu/intel_pstate/no_turbo
will not be present, in this case file not present
exception should be handled gracefully and print the
err message.

Signed-off-by: Reshma Pattan <reshma.pattan@intel.com>